+++ lib/mahttp.go
@@ -1,4 +1,4 @@
-package mahttp
+package lib
 
 import (
 	"net/url"
@@ -6,11 +6,11 @@ import (
 	"github.com/multiformats/go-multiaddr"
 )
 
-// JoinPath joins a HTTP path onto an existing multiaddr. If the multiaddr
+// JoinHTTPPath joins a HTTP path onto an existing multiaddr. If the multiaddr
 // includes the "http-path" protocol already, it is joined to the end of the
 // existing path. If the multiaddr does not contain the "http-path" protocol
 // then it is appended to the multiaddr, with the value specified.
-func JoinPath(addr multiaddr.Multiaddr, path string) (multiaddr.Multiaddr, error) {
+func JoinHTTPPath(addr multiaddr.Multiaddr, path string) (multiaddr.Multiaddr, error) {
 	found := false
 	var out multiaddr.Multiaddr
 	for _, comp := range addr {
